requirements.txt
pandas>=1.5.2
numpy>=1.24.1
python-dotenv>=0.21.0
psycopg2
django